Paycheck returns to original after restarting
For level two at least, when the player restarts the level. the paycheck returns to it's original pay.
For example, if the player's original paycheck is $570, and if the player restarts the level at the paycheck of $630 or $515, it'll return to $570.
I've also fixed the MP3 canvas to be behind of the background for level one.