Additions and a Change
I've added sound effects for the power ups when collecting them
I've change the button for the quit ending to L, because both that and the MP3 player uses the same button.
I've added a MP3 scene so the player can listen to all songs in the MP3 player without worrying about anything.